Correct type of table id

+++ include/zlimdbclient.h
@@ -41,7 +41,7 @@ int zlimdb_is_connected(zlimdb* zdb);
 int zlimdb_errno();
 const char* zlimdb_strerror(int errnum);
 
-int zlimdb_add_table(zlimdb* zdb, const char* name, uint64_t* table_id);
+int zlimdb_add_table(zlimdb* zdb, const char* name, uint32_t* table_id);
 
 int zlimdb_add(zlimdb* zdb, uint32_t table_id, const void* data, uint32_t size);
 int zlimdb_query(zlimdb* zdb, uint32_t table_id, zlimdb_query_type type, uint64_t param);
